Experiment defining structs with an env
Struct members need to know things like the target aligment or the size
of a pointer. Defining a struct member is now done via a function which
accepts an environment and returns the proper member. For instance,
primitive types which were previously concrete maps are now such
functions. The whole scheme allows for more code reuse and less
boilerplate.